Add initial Infrastucture preparations

+++ Backend/Incidents/Incidents.Domain/Entities/Contact.cs
@@ -4,9 +4,9 @@ using System.Text;
 
 namespace Incidents.Domain.Entities
 {
-    public class Contact : IEntity<int>
+    public class Contact : IEntity<long>
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Email { get; set; }
