+++ Backend/Incidents/Incidents.Domain/Entities/Incident.cs
@@ -4,9 +4,9 @@ using System.Text;
 
 namespace Incidents.Domain.Entities
 {
-    public class Incident : IEntity<int>
+    public class Incident : IEntity<long>
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string IncidentName { get; set; }
         public ICollection<Account> Accounts { get; set; }
     }
